+++ lib/cassandra/address_resolution/policies.rb
@@ -1,0 +1,2 @@
+require 'cassandra/address_resolution/policies/none'
+require 'cassandra/address_resolution/policies/ec2_multi_region'